+++ src/keys.c
@@ -10,17 +10,12 @@ meta* recognize_key(char key, meta* Dt)
 			free_all_exit(1, Dt);
 
 		default:
-			/* Only printable chars will be added. Combinations that aren't
-			specified above will be omited. Set "key != ESC" to enable. */
-			if(key == NUL || key == LF || key >= 32)
-			{
-				Dt = non_control_chr(key, Dt);
-			}
+			Dt = text_char(key, Dt);
 			break;
 
 		case HT:
 			// Currently converts the tab to one space.
-			Dt = non_control_chr(' ', Dt);
+			Dt = text_char(' ', Dt);
 			break;
 
 		case DEL:
@@ -45,34 +40,47 @@ meta* recognize_key(char key, meta* Dt)
 	return Dt;
 }
 
-meta* non_control_chr(char key, meta* Dt)
+meta* text_char(char key, meta* Dt)
 {
-	if(Dt->chars < BUF_MAX)
+	/* Only printable chars will be added. Combinations that aren't specified
+	above will be omited. Set "key != ESC" to enable. */
+	if((key == NUL || (key >= BEL && key <= FF) || key >= 32))
 	{
-		Dt->chars++;
-		ACT_LN_LEN++;
-
-		Dt = extend_act_ln_mem(Dt);
-
-		/* If the cursor is moved to the left and a char is inserted, rest of
-		the text will be shifted to the right side. */
-		if(Dt->cusr_x > 0)
+		if(Dt->chars < BUF_MAX)
 		{
-			Dt = shift_text_horizonally('r', Dt);
-		}
-		ACT_LN[ACT_LN_LEN - Dt->cusr_x - NUL_SZ] = key;
-		ACT_LN[ACT_LN_LEN] = NUL;
+			Dt->chars++;
+			ACT_LN_LEN++;
 
-		// Initializer handling.
-		if(key == NUL && ACT_LN_LEN > 0)
-		{
-			Dt->chars--;
-			ACT_LN_LEN--;
+			Dt = extend_act_line_mem(Dt);
+
+			/* If the cursor is moved to the left and a char is inserted, rest
+			of the text will be shifted to the right side. */
+			if(Dt->cusr_x > 0)
+			{
+				Dt = shift_text_horizonally('r', Dt);
+			}
+			ACT_LN[ACT_LN_LEN - Dt->cusr_x - NUL_SZ] = key;
+			ACT_LN[ACT_LN_LEN] = NUL;
+
+			// Initializer handling.
+			if(key == NUL && ACT_LN_LEN > 0)
+			{
+				Dt->chars--;
+				ACT_LN_LEN--;
+			}
+			else if(key == LF)
+			{
+				Dt = linefeed(Dt);
+			}
 		}
-		else if(key == LF)
+		else
 		{
-			Dt = linefeed(Dt);
+			SET_STATUS("can't read or insert more chars\0");
 		}
+	}
+	else
+	{
+		SET_STATUS("binary file, damage after saving\0");
 	}
 	return Dt;
 }
@@ -91,18 +99,18 @@ meta* linefeed(meta* Dt)
 		will be moved to the new line. */
 		if(Dt->cusr_x > 0)
 		{
-			for(term_t x = PREV_LN_LEN - Dt->cusr_x; x < PREV_LN_LEN; x++)
+			for(buf_t x = PREV_LN_LEN - Dt->cusr_x; x < PREV_LN_LEN; x++)
 			{
 				ACT_LN[ACT_LN_LEN] = PREV_LN[x];
 				ACT_LN_LEN++;
-				Dt = extend_act_ln_mem(Dt);
+				Dt = extend_act_line_mem(Dt);
 			}
 
 			// Now the length of the upper line will be shortened after copying.
 			PREV_LN_LEN -= Dt->cusr_x;
 			PREV_LN[PREV_LN_LEN] = NUL;
 
-			Dt = shrink_prev_ln_mem(Dt);
+			Dt = shrink_prev_line_mem(Dt);
 		}
 		ACT_LN[ACT_LN_LEN] = NUL;
 	}
@@ -118,7 +126,7 @@ meta* backspace(meta* Dt)
 		if(Dt->cusr_x != ACT_LN_LEN)
 		{
 			Dt = shift_text_horizonally('l', Dt);
-			Dt = shrink_act_ln_mem(Dt);
+			Dt = shrink_act_line_mem(Dt);
 
 			ACT_LN_LEN--;
 			Dt->chars--;
@@ -131,27 +139,12 @@ meta* backspace(meta* Dt)
 			for(buf_t x = 0; x <= ACT_LN_LEN; x++)
 			{
 				PREV_LN[PREV_LN_LEN] = ACT_LN[x];
+
 				if(ACT_LN[x] != NUL)
 				{
 					PREV_LN_LEN++;
 				}
-
-				if(PREV_LN_LEN == INIT_MEMBLK)
-				{
-					// If there are 4/8 chars + terminator, extend to MEMBLK.
-					PREV_LN = realloc(PREV_LN, MEMBLK);
-					puts("ALLOC_EIGHT");
-				}
-				else if(PREV_LN_LEN > INIT_MEMBLK && PREV_LN_LEN % MEMBLK == 0)
-				{
-					PREV_LN = realloc(PREV_LN,
-					((PREV_LN_LEN / MEMBLK) * MEMBLK) + MEMBLK);
-
-					printf("ALLOC: %d\n", ((PREV_LN_LEN / MEMBLK) * MEMBLK)
-					+ MEMBLK);
-				}
-				chk_ptr(PREV_LN, "extend a memblock for the actent line\0",
-				Dt);
+				Dt = extend_prev_line_mem(Dt);
 			}
 			free(ACT_LN);
 			ACT_LN = NULL;
@@ -167,7 +160,7 @@ meta* backspace(meta* Dt)
 		ACT_LN = NULL;
 		Dt->lines--;
 
-		Dt = shrink_act_ln_mem(Dt);
+		Dt = shrink_act_line_mem(Dt);
 
 		ACT_LN_LEN--;
 		Dt->chars--;