+++ doc/CHANGELOG.md
@@ -8,6 +8,8 @@ files use git-show or git-log.
 
 ## Versions:
 ### v2.1.0 (WIP):
+- new file limit: 16 Mb - 1,
+- messages in the upper bar,
 - improved Makefile,
 - more efficent memory allocation.
 - control keys changed,